+++ project/taskmanager/storage.go
@@ -18,13 +18,13 @@ type Storage interface {
 	// List returns every stored task.
 	List(ctx context.Context) ([]Task, error)
 	// Get returns the task with the given identifier, or ErrTaskNotFound.
-	Get(ctx context.Context, id int) (Task, error)
+	Get(ctx context.Context, id int64) (Task, error)
 	// Add stores a new task with the given title and returns it with an
 	// assigned identifier.
 	Add(ctx context.Context, title string) (Task, error)
 	// Complete marks the task with the given identifier as done and returns
 	// the updated task, or ErrTaskNotFound.
-	Complete(ctx context.Context, id int) (Task, error)
+	Complete(ctx context.Context, id int64) (Task, error)
 	// Remove deletes the task with the given identifier, or ErrTaskNotFound.
-	Remove(ctx context.Context, id int) error
+	Remove(ctx context.Context, id int64) error
 }
